move src/utils/getViewsProperties.js to src/utils/getViewsProperties.js
